models: First stage of rewriting Event objects into Django ORM. - This commit breaks everything.   - Stage 2 will wire everything back together.
